Revert "relocate utils dbus code to enable sharing"
This reverts commit 0559ace4087244d41747008aa2e0473e92bfdc23.
The requirement for this code in bluealsa-client was removed by
commit 1d92201 when the monitor command was re-written.
